Move Icon component into a package
