+++ packages/icon/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@hig/icon",
+  "version": "0.1.0-alpha",
+  "description": "HIG Icon component",
+  "author": "Autodesk Inc.",
+  "license": "Apache-2.0",
+  "homepage": "https://hig.autodesk.com",
+  "main": "build/index.js",
+  "module": "build/index.es.js",
+  "dependendies": {
+    "@hig/icons": "^0.1.0-alpha"
+  },
+  "devDependencies": {
+    "@hig/scripts": "^0.1.0",
+    "@hig/styles": "^0.1.0-alpha"
+  },
+  "peerDependencies": {
+    "react": "^15.4.1",
+    "prop-types": "^15.6.1"
+  },
+  "scripts": {
+    "build": "hig-scripts-build"
+  }
+}
