+++ packages/icon/src/Icon.test.js
