+++ packages/icon/src/availableSizes.js
@@ -1,0 +1,6 @@
+import sizes from './sizes';
+
+/** @type {string[]} */
+const availableSizes = Object.values(sizes);
+
+export default availableSizes;
