+++ packages/icon/src/icon.scss
