+++ packages/icon/src/index.js
@@ -1,0 +1,1 @@
+export { default, AVAILABLE_NAMES, names, sizes } from './Icon';
