+++ packages/icon/src/sizes.js
@@ -1,0 +1,1 @@
+export default Object.freeze({ PX_24: "24", PX_16: "16" });
