+++ packages/react/package.json
@@ -28,6 +28,7 @@
     "type-check": "tsc"
   },
   "dependencies": {
+    "@hig/icon": "0.1.0-alpha",
     "@hig/icons": "0.1.0-alpha",
     "hig-vanilla": "^0.1.26",
     "react-flip-move": "^3.0.1",
